PylectroRAT/requirements.txt
flask
pyautogui
pynput
pycryptodome
